requirements.txt
# requirements.txt
crewai[tools]==0.40.0
litellm==1.34.0
matplotlib==3.7.2
pandas==2.1.0
requests==2.31.0
beautifulsoup4==4.12.2
plotly==5.15.0
python-dotenv==1.0.0
seaborn==0.12.2
numpy==1.24.3
yfinance==0.2.18
tweepy==4.14.0
youtube-transcript-api==0.6.1
textblob==0.17.1
vaderSentiment==3.3.2
feedparser==6.0.10
retry==0.9.2